tests/wpt/tests/html/canvas/element/pixel-manipulation/2d.imageData.put.path.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.imageData.put.path</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.imageData.put.path</h1>
<p class="desc">putImageData() does not affect the current path</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("putImageData() does not affect the current path");
_addTest(function(canvas, ctx) {

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50)
  ctx.rect(0, 0, 100, 50);
  var imgdata = ctx.getImageData(0, 0, 100, 50);
  ctx.putImageData(imgdata, 0, 0);
  ctx.fillStyle = '#0f0';
  ctx.fill();
  _assertPixelApprox(canvas, 50,25, 0,255,0,255, 2);

});
</script>
